feat: Modify the authentication.
Use the deepin-deb-installer to install the software for the first time must be authenticated when deepin-deb-installer opened every time. If the interval between later installation of the software package and the first installation dose not exceed the set time, authentication is not required. After the time has expired, authenticate again and update the authentication time.

Log: Modify the authentication.
Change-Id: Ic0deb6959930a1049090c97f1caba299bdce4628